third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-events.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup events</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>

<div popup>Popup</div>

<script>
promise_test(async t => {
  const popup = document.querySelector('[popup]');
  let showCount = 0;
  let hideCount = 0;
  await new Promise(resolve => window.addEventListener('load',() => resolve()));
  assert_false(popup.matches(':top-layer'));
  document.addEventListener('show',() => ++showCount);
  document.addEventListener('hide',() => ++hideCount);
  assert_equals(0,showCount);
  assert_equals(0,hideCount);
  popup.showPopUp();
  assert_true(popup.matches(':top-layer'));
  assert_equals(1,showCount);
  assert_equals(0,hideCount);
  await waitForRender();
  assert_true(popup.matches(':top-layer'));
  popup.hidePopUp();
  assert_false(popup.matches(':top-layer'));
  assert_equals(1,showCount);
  assert_equals(1,hideCount);
  await waitForRender();
  // No additional events after animation frame
  assert_false(popup.matches(':top-layer'));
  assert_equals(1,showCount);
  assert_equals(1,hideCount);
}, 'Show and hide events get properly dispatched for popups');

</script>
